# ENH: Improve performance of np.broadcast_arrays and np.broadcast_shapes