@seberg 2024-04-04
> fine that arrays are only made read-only if their shape is actually changes

Just saw the emails and hadn't noticed this, so a brief chiming in.  I am on edge about "unpredictable" readonly flag: That is the exact kind of things that the tuple vs. list return discussion was about, although I don't know if e.g. `numba` encodes it (or we might ever encode it in static typing).
It seems fine and is a pretty small thing but I also don't think there is any up-side.  If a user knows that it doesn't change, `np.broadcast_to` with the unchanged array shape makes more sense.

For `nditer` to set this you need to pass `op_flags=[["writeonly"]]*len(arrs)` or so, probably.  TBH, it may also make sense to move `broadcast_to`, `broadcast_shapes` or something similar to C instead (with the goal of reusing at least the core broadcasting code in `nditer` one day...  (the code there can use some love, IMO)

(The magic `nargs` "if" should be `<=64` now.)