@eendebakpt 2024-04-04
@mhvk @seberg For comparison I created https://github.com/numpy/numpy/pull/26214 which uses `np.nditer` to be perform the broadcasting. That PR makes all output of `np.broadcast_arrays` readonly, but I cannot judge wether that is acceptable (now or after some longer deprecation period).