@mhvk 2024-05-31
Overall, I think that this change is a nice simple one that doesn't really affect much, so would suggest to put it in. We can perhaps separately discuss #26214, of whether we should move to `nditer`? @seberg - what do you think?