@seberg 2024-06-02
OK, fair, I guss sometimes it was alrady not readonly.  Although I do wonder if creating the view always wouldn't be better.  But thanks @eendebakpt and @mhvk.